Refactor job post form schemas and imports for better organization
